Implementando esquema de base de datos no relacional

+++ src/main/java/com/nttdata/bootcamp_customer_service/model/collection/BankProduct.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Document(collection = "bank_products")
@@ -16,7 +17,7 @@ public class BankProduct {
     private String productNumber;  // Unique product number
     private String typeProductId;  // Reference to the product type (ProductType)
     private boolean isActive;  // Product status (active or not)
-    private double balance;  // Product balance
+    private BigDecimal balance;  // Product balance
 
     private List<AccountHolder> accountHolders;  // List of holders (titulares) for this product
     private List<Transaction> transactions;  // List of transactions (movements) for this product
